DAY 3 - TIP CALCULATOR
A good project for learners to practice fundamental arithmetic operations and user input handling in Python is to create a tip calculator. To begin, ask the user to enter the total amount of the bill and the percentage they would like to tip. Next, multiply the bill amount by the tip % and divide the result by 100 to determine the tip amount. To determine the final amount, add this tip amount to the total bill. If you'd like, you may arrange the output to show the tip amount, the total bill with the tip included, or even divide the bill among a certain number of recipients. With Python's input function, you can quickly create a useful tip calculator software that allows for user involvement and simple arithmetic operations.